+++ src/model/Promocion.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface Promocion {
 	
-	public int aplicarDescuento(List<Atraccion> atracciones);
+	public float aplicarDescuento(List<Atraccion> atracciones);
 }
